strip out configuration to its own class to make pycryptobot class more readable
